+++ test/New-DelimitedText.Tests.ps1
@@ -1,7 +1,7 @@
 #Requires -Modules @{ ModuleName='Pester'; ModuleVersion='5.0.0' }, PsDac
 
 Describe New-DelimitedText {
-    
+
     BeforeAll {
         Import-Module $PSScriptRoot\..\src\PsDataFactory.psd1 -Force -ErrorAction Stop
     }
@@ -11,6 +11,18 @@ Describe New-DelimitedText {
     }
 
     It works {
-        New-AdfDelimitedTextDataset -Name MyDataset -TableName MyTable -LinkedServiceReference $LinkedServiceReference -ErrorAction Stop
+        New-AdfDelimitedTextDataset `
+            -Name MyDataset `
+            -LinkedServiceReference $LinkedServiceReference `
+            -Location ([PSCustomObject]@{
+                type = "type"
+                folderPath = "folderPath"
+                container = "contain"
+            }) `
+            -ColumnDelimiter "`t" `
+            -EscapeChar "`\" `
+            -FirstRowAsHeader $true `
+            -QuoteChar "`"" `
+            -ErrorAction Stop
     }
 }